+++ Source/Jeeves/Program.cs
@@ -1,16 +1,20 @@
 ﻿using System;
+using System.IO;
+using Jeeves.Core;
 using Serilog;
+using Topshelf;
 
 namespace Jeeves
 {
     public static class Program
     {
-        private const string SettingsPath = @"C:\ProgramData\Jeeves\settings.json";
-        private const string Database = @"C:\ProgramData\Jeeves\Jeeves.sqlite";
-        private const string SqlScritpsDirectory = @"C:\ProgramData\Jeeves\Scripts";
-        private const string LogFile = @"C:\ProgramData\Jeeves\Logs\{Date}.txt";
+        private const string JeevesPath = @"C:\ProgramData\Jeeves\";
+        private const string SettingsPath = JeevesPath + "settings.json";
+        private const string Database = JeevesPath + "Jeeves.sqlite";
+        private const string SaltPath = JeevesPath + "salt";
+        private const string LogFile = JeevesPath + @"Logs\{Date}.txt";
 
-        public static void Main(string[] args)
+        public static void Main()
         {
             Log.Logger = new LoggerConfiguration()
                 .MinimumLevel.Debug()
@@ -20,19 +24,39 @@ namespace Jeeves
 
             try
             {
-                var settings = JeevesSettingsLoader.Load(SettingsPath);
+                if (!Directory.Exists(JeevesPath))
+                {
+                    Directory.CreateDirectory(JeevesPath);
+                }
 
-                Topshelf.Start(Database, SqlScritpsDirectory, settings);
+                var settings = JeevesSettingsLoader.Load(SettingsPath);
+                var salt = Hasher.LoadSalt(SaltPath);
+                var store = new SQLiteStore(Database, salt);
+
+                HostFactory.Run(hc =>
+                {
+                    hc.Service<JeevesHost>(sc =>
+                    {
+                        sc.ConstructUsing(() => new JeevesHost(settings, store, new JeevesLog()));
+                        sc.WhenStarted(s => s.Start());
+                        sc.WhenStopped(s =>
+                        {
+                            s.Stop();
+                            s.Dispose();
+                        });
+                    });
+
+                    hc.SetDescription("A simple REST service which provides configuration data for applications");
+                    hc.SetServiceName("Jeeves");
+                });
+
+                Log.CloseAndFlush();
             }
             catch (Exception e)
             {
                 Log.Error(e, "Could not start Jeeves");
                 Log.CloseAndFlush();
                 Environment.Exit(-1);
-            }
-            finally
-            {
-                Log.CloseAndFlush();
             }
         }
     }